layout: Scroll-bar shift bad fix
The scroll-bar caused a layout shift between pages that had a scroll bar
and pages that did not. This was nasty and super noticeable because of
the margin lines component. This change updates the global.css file to
get rid of the scrollbar entirely for now. This may not be the approved
solution but it will do for now